+++ crates/base/src/server.rs
@@ -2,12 +2,12 @@ use crate::worker_ctx::{
     create_events_worker, create_user_worker_pool, create_worker, WorkerRequestMsg,
 };
 use anyhow::Error;
+use event_manager::events::WorkerEventWithMetadata;
 use hyper::{server::conn::Http, service::Service, Body, Request, Response};
 use log::{debug, error, info};
 use sb_worker_context::essentials::{
     MainWorkerRuntimeOpts, WorkerContextInitOpts, WorkerRuntimeOpts,
 };
-use sb_worker_context::events::WorkerEventWithMetadata;
 use std::future::Future;
 use std::net::IpAddr;
 use std::net::Ipv4Addr;
