+++ crates/event_manager/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sb_worker_context"
+name = "event_manager"
 version = "0.1.0"
 authors = ["Supabase <team@supabase.com>"]
 edition = "2021"
@@ -11,10 +11,9 @@ license = "MIT"
 path = "lib.rs"
 
 [dependencies]
-anyhow = { workspace = true }
-enum-as-inner = "0.6.0"
-hyper.workspace = true
-serde.workspace = true
-tokio.workspace = true
+deno_core.workspace = true
 uuid.workspace = true
-event_manager = { version = "0.1.0", path = "../event_manager" }
+serde.workspace = true
+anyhow.workspace = true
+tokio.workspace = true
+log.workspace = true
