+++ crates/event_manager/event_worker.js
