+++ crates/sb_worker_context/lib.rs
@@ -1,2 +1,1 @@
 pub mod essentials;
-pub mod events;
